# web/openapi.json
{
  "components": {
    "schemas": {
      "ActiveDefault": {
        "description": "One setting this server turns on even when the run would have fitted without\nit.\n\nThe list is the *whole* of what invariant 4 now permits without an opt-in: a\ndegradation not named here is still a 422 that asks for `allow`.",
        "properties": {
          "client_can_disable": {
            "description": "Always true: one `params` field turns it off.",
            "type": "boolean"
          },
          "condition": {
            "description": "When it is applied, in one sentence.",
            "type": "string"
          },
          "cost": {
            "description": "What the run pays for it.",
            "type": "string"
          },
          "degrades": {
            "description": "Whether leaving it on changes *what* is computed rather than how fast.",
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "note": {
            "description": "What it does.",
            "type": "string"
          },
          "paths": {
            "description": "The dotted paths it moves. Setting any of them in `params` pins it.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "warning": {
            "description": "The `plan.warnings` code this default raises, when it raises one.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "condition",
          "cost",
          "note",
          "paths",
          "degrades",
          "client_can_disable"
        ],
        "type": "object"
      },
...
      "Defaults": {
        "description": "`GET /v1/defaults`",
        "properties": {
          "active_defaults": {
            "items": {
              "$ref": "#/components/schemas/ActiveDefault"
            },
            "type": "array"
          },
          "derived": {
            "items": {
              "$ref": "#/components/schemas/DerivedField"
            },
            "type": "array"
          }
        },
        "required": [
          "derived",
          "active_defaults"
        ],
        "type": "object"
      },
      "DerivedField": {
        "description": "One field the resolver decides when `params` leaves it out, and the rule it\ndecides it with.\n\nThis is what replaced `GET /v1/presets`. A profile was a word the server had\nto interpret; this is the interpretation itself, published - a client that\nwants its own \"fast\" or \"thorough\" builds one out of these rules and sends it\nas `params`.",
        "properties": {
          "applies_to": {
            "description": "`any` | `sft` | `grpo` | `ppo` | `rollout` | `evaluation` | `checkpoint`:\nwhen the rule applies at all.",
            "type": "string"
          },
          "client_can_set": {
            "description": "Always true, and stated rather than implied: every one of these is a\n`params` field a client may set instead.",
            "type": "boolean"
          },
          "path": {
            "description": "Dotted field path, the same grammar as `params`, `provenance` keys and\nthe `PATCH` whitelist. One path grammar for the whole API.",
            "type": "string"
          },
          "rule": {
            "description": "The rule, in one sentence.",
            "type": "string"
          },
          "thresholds": {
            "additionalProperties": {
              "format": "double",
              "type": "number"
            },
            "description": "The numbers in that sentence, named. A `BTreeMap` so the bytes are\ndeterministic.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "path",
          "rule",
          "applies_to",
          "client_can_set"
        ],
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "The closed vocabulary a client's `match` is written against.\n\n`message` and `hint` are prose and may be reworded release to release;\n`code` is the part of the contract, tested (`tests/error_catalog.rs`)\nagainst `ERRORS.md`.\n\nThe spelling beside each variant is the one serde writes *and* the one\n`as_str()` returns - one token, so the two cannot drift and no test has\nto check them against each other.\n\n`rename_all` is for the schema, which reads the container rule and not\nthe per-variant spelling; every spelling below is already snake case.",
        "enum": [
          "missing_field",
          "unknown_field",
          "invalid_value",
          "out_of_range",
          "unsupported_format",
          "path_not_found",
          "forbidden_path",
          "unknown_catalog_id",
          "needs_opt_in",
          "override_conflict",
          "conflict",
          "unauthorized",
          "not_found",
          "server_declared"
        ],
        "type": "string"
      },
...
      "FieldError": {
        "description": "One field-level problem, located by a JSON Pointer (RFC 6901) into the\nrequest body.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "hint": {
            "description": "What to do about it, only when there is a short, concrete answer. A\nvague hint is worse than none.",
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "pointer": {
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ModelFile": {
        "description": "One GGUF file under the path roots.",
        "properties": {
          "bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "modified_at": {
            "description": "Unix seconds.",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "path": {
            "description": "Exactly what `recipe.model` accepts.",
            "type": "string"
          },
          "relative": {
            "description": "`path` relative to `root`, for display.",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ModelFileRole"
          },
          "root": {
            "description": "The root it was found under.",
            "type": "string"
          }
        },
        "required": [
          "path",
          "root",
          "relative",
          "bytes",
          "role"
        ],
        "type": "object"
      },
      "ModelFileListing": {
        "description": "`GET /v1/model-files`",
        "properties": {
          "files": {
            "description": "Sorted by path.",
            "items": {
              "$ref": "#/components/schemas/ModelFile"
            },
            "type": "array"
          },
          "roots": {
            "description": "The roots walked: the server's `path_roots`, or its working directory\nwhen it declares none.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "truncated": {
            "description": "The walk stopped at its bound; some files are not listed.",
            "type": "boolean"
          }
        },
        "required": [
          "roots",
          "files",
          "truncated"
        ],
        "type": "object"
      },
      "ModelFileRole": {
        "description": "What a GGUF file is for, judged by its name.",
        "enum": [
          "model",
          "projector",
          "vocab"
        ],
        "type": "string"
      },
...
      "PreflightRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/preflight`",
        "properties": {
          "device": {
            "description": "`auto` | `cpu` | `gpu`; defaults to `auto`.",
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "targets": {
            "description": "LoRA target shorthands or patterns, as the CLI's `--targets` accepts.\nAbsent selects the architecture's automatic profile.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "model"
        ],
        "type": "object"
      },
      "PreflightResponse": {
        "allOf": [
          {
            "description": "Flattened so the HTTP response is the versioned engine contract itself,\nnot an envelope carrying a second schema.",
            "type": "object"
          }
        ]
      },
      "Problem": {
        "description": "Wire shape of a problem document (RFC 9457). Serialized by hand rather than\nderived from `ApiError` so the field order - and therefore the bytes - is\nfixed. Published as the `Problem` component: every failure of the API but\nthe OpenAI routes' has this body.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "meta": {
            "description": "The payload specific to the problem type, e.g. the dominant memory\nposts of `insufficient-memory`.",
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "description": "The id every log line of this request carries.",
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "description": "`https://retrograd.dev/problems/<slug>`: the slug is what a client\nbranches on.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP control plane for LoRA training runs.",
    "title": "retrograd-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/defaults": {
      "get": {
        "operationId": "get_defaults",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Defaults"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Derivation rules and the settings applied by default"
      }
    },
...
    "/v1/model-files": {
      "get": {
        "operationId": "get_model_files",
        "parameters": [
          {
            "description": "Read the disk again instead of the listing of the last 30 seconds.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelFileListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "GGUF files under the path roots, by name, for picking a model"
      }
    },
...
    "/v1/preflight": {
      "post": {
        "operationId": "post_preflight",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Build the training graph and report on it"
      }
    },
...
  }
}